% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewRetinaSC.R
\name{viewRetinaSC}
\alias{viewRetinaSC}
\title{View HumanLiver data in the scClustViz Shiny app}
\usage{
viewRetinaSC(dataSet, outPath = "./")
}
\arguments{
\item{dataSet}{One of \code{"All"} or \code{"RSCsOnly"} for the full dataset
or the subset of predicted retinal stem cell clusters respectively.}

\item{outPath}{Default = "./" (the working directory). Specify the directory
used to save/load any analysis files you generate while exploring the
\code{HumanLiver} data.}
}
\value{
The function causes the scClustViz Shiny GUI app to open in a
  seperate window.
}
\description{
A wrapper function to view the \code{HumanLiver} dataset in the
\code{scClustViz} Shiny app.
}
\examples{
\dontrun{
  viewRetinaSC("All")
}

}
\seealso{
\url{https://baderlab.github.io/scClustViz} for information on
  \code{scClustViz}.
}
