% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewRetinaSC.R
\name{viewRetinaSC}
\alias{viewRetinaSC}
\title{View HumanLiver data in the scClustViz Shiny app}
\usage{
viewRetinaSC(dataSet = "All", outPath = "./", imageFileType = "tiff", ...)
}
\arguments{
\item{dataSet}{One of \code{"All"} or \code{"RSCsOnly"} for the full dataset
or the subset of predicted retinal stem cell clusters respectively.}

\item{outPath}{Default = "./" (the working directory). Specify the directory
used to save/load any analysis files you generate while exploring the
\code{HumanLiver} data.}

\item{imageFileType}{Default="pdf". The file format for saved figures. One of
\code{"pdf"} (generated with \code{\link[grDevices]{cairo_pdf}}),
\code{"eps"} (generated with \code{\link[grDevices]{cairo_ps}}),
\code{"tiff"} (generated with \code{\link[grDevices]{tiff}}), or
\code{"png"} (generated with \code{\link[grDevices]{png}}).}

\item{...}{Named options that should be passed to the
\code{\link[shiny]{runApp}} call (these can be any of the following:
"port", "launch.browser", "host", "quiet", "display.mode" and "test.mode").}
}
\value{
The function causes the scClustViz Shiny GUI app to open in a
  seperate window.
}
\description{
A wrapper function to view the \code{HumanLiver} dataset in the
\code{scClustViz} Shiny app.
}
\examples{
\dontrun{
  viewRetinaSC("All")
}

}
\seealso{
\url{https://baderlab.github.io/scClustViz} for information on
  \code{scClustViz}.
}
